Set PIN reenter value check
