+++ ni_sdk/src/main/java/ae/network/nicardmanagementsdk/presentation/ui/set_pin/SetPinFragment.kt
@@ -37,6 +37,7 @@ abstract class SetPinFragment : SetPinDialogFragmentBase<SetPinViewModel>() {
 
     override fun initializeUI() {
         super.initializeUI()
+
         viewModel.onResultSingleLiveEvent.observe(this) { successErrorResponse ->
             successErrorResponse?.let { response ->
                 lifecycleScope.launch {
