--- ni_sdk/src/main/java/ae/network/nicardmanagementsdk/presentation/ui/set_pin/SetVerifyPinBaseViewModel.kt
@@ -1,25 +1,0 @@
-package ae.network.nicardmanagementsdk.presentation.ui.set_pin
-
-import ae.network.nicardmanagementsdk.core.ISetVerifyPinCore
-import ae.network.nicardmanagementsdk.network.utils.ConnectionModel
-import ae.network.nicardmanagementsdk.network.utils.IConnection
-import androidx.lifecycle.viewModelScope
-import kotlinx.coroutines.launch
-
-abstract class SetVerifyPinBaseViewModel (
-    private val setVerifyPinCore: ISetVerifyPinCore,
-    private val connectionLiveData: IConnection<ConnectionModel>
-) : SetPinViewModelBase() {
-
-    override fun onDoneImageButtonTap() {
-        viewModelScope.launch {
-            if (connectionLiveData.hasInternetConnectivity) {
-                isVisibleProgressBar.value = true
-                val result = setVerifyPinCore.makeNetworkRequest(inputString)
-                isVisibleProgressBar.value = false
-                onResultSingleLiveEvent.value = result
-            }
-        }
-    }
-
-}
