+++ ni_sdk/src/main/java/ae/network/nicardmanagementsdk/presentation/ui/verify_pin/VerifyPinViewModel.kt
@@ -4,14 +4,27 @@ import ae.network.nicardmanagementsdk.R
 import ae.network.nicardmanagementsdk.core.ISetVerifyPinCore
 import ae.network.nicardmanagementsdk.network.utils.ConnectionModel
 import ae.network.nicardmanagementsdk.network.utils.IConnection
-import ae.network.nicardmanagementsdk.presentation.ui.set_pin.SetVerifyPinBaseViewModel
+import ae.network.nicardmanagementsdk.presentation.ui.set_pin.SetPinViewModelBase
 import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.viewModelScope
+import kotlinx.coroutines.launch
 
 class VerifyPinViewModel (
-    setVerifyPinCore: ISetVerifyPinCore,
-    connectionLiveData: IConnection<ConnectionModel>
-) : SetVerifyPinBaseViewModel(setVerifyPinCore, connectionLiveData) {
+    private val setVerifyPinCore: ISetVerifyPinCore,
+    private val connectionLiveData: IConnection<ConnectionModel>
+) : SetPinViewModelBase() {
 
     override val navTitle = MutableLiveData(R.string.verify_pin_title)
     override val screenTitle = MutableLiveData(R.string.verify_pin_description)
+
+    override fun onDoneImageButtonTap() {
+        viewModelScope.launch {
+            if (connectionLiveData.hasInternetConnectivity) {
+                isVisibleProgressBar.value = true
+                val result = setVerifyPinCore.makeNetworkRequest(inputString)
+                isVisibleProgressBar.value = false
+                onResultSingleLiveEvent.value = result
+            }
+        }
+    }
 }
