+++ ni_sdk/src/main/res/layout/activity_set_pin.xml
@@ -50,7 +50,7 @@
                 app:layout_constraintHorizontal_bias="0.5"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@+id/customBackNavigationView"
-                tools:text="@string/set_pin_description" />
+                tools:text="@string/set_pin_description_enter_pin" />
 
             <com.google.android.material.button.MaterialButton
                 android:id="@+id/oneButton"
