+++ ni_sdk/src/main/res/values-ar/strings.xml
@@ -1,10 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    // app name
-    <string name="app_name">NICardManagementSDK</string>
-
     // card details activity
-    <string name="card_details_title">Card details</string>
+    <string name="card_details_title">معلومات البطاقة</string>
 
     // card details fragment
     <string name="card_number">رقم البطاقة</string>
@@ -15,7 +12,9 @@
 
     // set pin screen
     <string name="set_pin_title">تعيين الرقم السري (PIN)</string>
-    <string name="set_pin_description">الرجاء ادخال الرقم السري للبطاقة (PIN)</string>
+    <string name="set_pin_description_enter_pin">الرجاء ادخال الرقم السري للبطاقة (PIN)</string>
+    <string name="set_pin_description_re_enter_pin">يرجى إعادة إدخال رقم السري لبطاقتك (PIN)</string>
+    <string name="set_pin_description_pin_not_match">الرقم السري لبطاقتك غير متطابق\nيرجى إعادة إدخال رقم السري لبطاقتك (PIN)</string>
 
     // verify pin screen
     <string name="verify_pin_title">"التحقق من رقم السري لبطاقتك (PIN)"</string>
@@ -23,6 +22,6 @@
 
     // change pin screen
     <string name="change_pin_title">تغيير الرقم السري (PIN)</string>
-    <string name="change_pin_description_step1">الرجاء ادخال الرقم السري الحالي (PIN)</string>
-    <string name="change_pin_description_step2">الرجاء ادخال الرقم السري الجديد (PIN)</string>
+    <string name="change_pin_description_enter_current_pin">الرجاء ادخال الرقم السري الحالي (PIN)</string>
+    <string name="change_pin_description_enter_new_pin">الرجاء ادخال الرقم السري الجديد (PIN)</string>
 </resources>
